+++ src/app/routes/__init__.py
@@ -1,5 +1,5 @@
 from .browser import browser
-from .config_editing import config_handler
+from .config_handler import config_handler
 from .runner import runner
 
 __all__ = ["browser", "config_handler","runner"]
