+++ src/app/templates/runner.html
@@ -13,7 +13,7 @@
         <a class="btn btn-outline-secondary" href="{{ url_for('index') }}">Terug</a>
     </div>
 
-    {% include "_modal.html" %}
+    {% include "_modal_continue_run.html" %}
 {% endblock %}
 
 {% block scripts %}
